refactor: split out multi-runner from main
